+++ Utilities/DebugConsole/changelog.md
@@ -2,7 +2,8 @@ Changelog of Debug Console, a Phoenix Point Mod by Sheepy
 
 # Version 3.1
 
-* New: Api command to dump gui tree
+* New: Api command to dump gui tree.
+* Fix: MethodInfo result from API will now be displayed by ToString.
 
 # Version 3, 2020-04-11