+++ Utilities/DebugConsole/changelog.md
@@ -3,7 +3,7 @@ Changelog of Debug Console, a Phoenix Point Mod by Sheepy
 # Version 3.1
 
 * New: Api command "zy.gui.dump" to dump gui tree of given GameObject / Transform to modnix loader log.
-* Fix: MethodInfo result from API will now be displayed by ToString.
+* Fix: MethodInfo result from API are now displayed by ToString.
 
 # Version 3, 2020-04-11